Static cast conversions from uint64_t to uint32_t

+++ frontend/common/AdminCmd.cpp
@@ -430,7 +430,7 @@ void AdminCmd::processArchiveRoute_Add(xrd::Response& response) {
   auto& tapepool = getRequired(OptionString::TAPE_POOL);
   auto& comment  = getRequired(OptionString::COMMENT);
 
-  m_catalogue.ArchiveRoute()->createArchiveRoute(m_cliIdentity, scn, cn, tapepool, comment);
+  m_catalogue.ArchiveRoute()->createArchiveRoute(m_cliIdentity, scn, static_cast<uint32_t>(cn), tapepool, comment);
 
   response.set_type(xrd::Response::RSP_SUCCESS);
 }
@@ -442,10 +442,10 @@ void AdminCmd::processArchiveRoute_Ch(xrd::Response& response) {
   auto& cn       = getRequired(OptionUInt64::COPY_NUMBER);
 
   if(auto comment = getOptional(OptionString::COMMENT); comment) {
-    m_catalogue.ArchiveRoute()->modifyArchiveRouteComment(m_cliIdentity, scn, cn, comment.value());
+    m_catalogue.ArchiveRoute()->modifyArchiveRouteComment(m_cliIdentity, scn, static_cast<uint32_t>cn, comment.value());
   }
   if(auto tapepool = getOptional(OptionString::TAPE_POOL); tapepool) {
-    m_catalogue.ArchiveRoute()->modifyArchiveRouteTapePoolName(m_cliIdentity, scn, cn, tapepool.value());
+    m_catalogue.ArchiveRoute()->modifyArchiveRouteTapePoolName(m_cliIdentity, scn, static_cast<uint32_t>cn, tapepool.value());
   }
 
   response.set_type(xrd::Response::RSP_SUCCESS);
@@ -457,7 +457,7 @@ void AdminCmd::processArchiveRoute_Rm(xrd::Response& response) {
   auto& scn = getRequired(OptionString::STORAGE_CLASS);
   auto& cn  = getRequired(OptionUInt64::COPY_NUMBER);
 
-  m_catalogue.ArchiveRoute()->deleteArchiveRoute(scn, cn);
+  m_catalogue.ArchiveRoute()->deleteArchiveRoute(scn, static_cast<uint32_t>cn);
 
   response.set_type(xrd::Response::RSP_SUCCESS);
 }
